refactor: move CustomInstaller to its own module in the root
